chore: update Next.js version to 15.3.2 and remove unused ThemeProvider component
feat: add AboutMeCode component to display coding statistics from WakaTime and GitHub

feat: implement syntax highlighting components for displaying code snippets

feat: create a footer component for the application

feat: develop NowPlayingWidget to show currently playing Spotify track with progress

refactor: create a new ThemeProvider component for theme management
