+++ app/about/components/code_syntax/syntax-highlighting-component.tsx
