+++ package.json
@@ -28,7 +28,7 @@
         "clsx": "^2.1.1",
         "eslint-config-next": "15.2.3",
         "lucide-react": "^0.501.0",
-        "next": "^15.2.4",
+        "next": "^15.3.2",
         "next-themes": "^0.3.0",
         "react": "18.3.1",
         "react-dom": "18.3.1",
